move CachedMessagesCallStack into its own file
